untested rover domain split from environemnt class
